🐞(fix) 권한 알럿 시점 문제 수정
